moved create todo mutation in a separate hook
