+++ src/modules/todo/components/todo/Todo.tsx
@@ -1,7 +1,8 @@
 import { useForm } from 'react-hook-form'
 import * as classes from './Todo.module.scss'
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
+import { useQuery } from '@tanstack/react-query'
 import { TodoApiService } from '../../api'
+import { useMutationToPostTodo } from '../../hooks/useMutationToPostTodo'
 
 type AddTodoFormValues = {
   text: string
@@ -10,23 +11,12 @@ type AddTodoFormValues = {
 function Todo() {
   const { register, handleSubmit, reset } = useForm<AddTodoFormValues>()
 
-  const queryClient = useQueryClient()
-
   const { isLoading, data } = useQuery({
     queryKey: ['todos'],
     queryFn: TodoApiService.retrieve,
   })
 
-  const mutation = useMutation({
-    mutationFn: TodoApiService.create,
-    onSuccess: () => {
-      queryClient
-        .invalidateQueries({ queryKey: ['todos'] })
-        .catch((err: unknown) => {
-          console.log(err)
-        })
-    },
-  })
+  const mutation = useMutationToPostTodo()
 
   const onSubmit = handleSubmit(({ text }) => {
     mutation.mutate({
